compatible with VS2017 and VS 2015
