+++ VSOStatusInspector.Package/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="HtmlAgilityPack" version="1.4.9" targetFramework="net45" />
+  <package id="HtmlAgilityPack" version="1.5.1" targetFramework="net45" />
+  <package id="Microsoft.VSSDK.BuildTools" version="15.0.26201" targetFramework="net45" developmentDependency="true" />
 </packages>
